Use bindgen to generate alsa-sys bindings
While working on #363 I noticed I was missing a few items from the ALSA
lib. Rather than adding them by hand, I thought I would update the crate
using the latest version of bindgen. I'm on the NixOS unstable channel
and have used the `alsaLib.dev` package headers as the source for
bindings generation.

I noticed that bindgen also generated the parts of `libc` that we were
using, so I removed the `libc` dep in favour of the generated bindings.
I'm not sure what the best-practise approach is here - that is, whether
it is better to use `libc` where possible, or to use what bindgen gives
us. rust-bindgen itself doesn't give us an option to use `libc`, so I
think you'd have to use some extensive black/white-listing to get it
working.

There are a few warnings in the generated bindings related to u128 types
having an unstable ABI. This seems to be a bug in
rust-lang/rust-bindgen#1560. Rather than fixing this by hand I'd prefer
to regenerate the bindings once the bug has been fixed.